Add CMakeLists.txt for pricing_app_updater
A new CMakeLists.txt file has been added to manage the build process of the 'pricing_app_updater' project. It specifies the minimum required version of CMake, defines the project name, and the standards for C++. It also sets the output directories for executables and debug, and names the executable as 'updater'.